add search enpoint into references sidebar
