+++ src/components/map/Sidebar/SidebarV2/SearcherSidebar.tsx
@@ -64,7 +64,7 @@ type SearcherSidebarProps = {
   preLoadNodes: (nodeIds: string[], fullNodes: FullNodeData[]) => Promise<void>;
 };
 
-type Pagination = {
+export type Pagination = {
   data: any[];
   lastPageLoaded: number;
   totalPage: number;
